requirements.txt
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.6.2.post1
asyncpg==0.30.0
bidict==0.23.1
blinker==1.8.2
cffi==1.17.1
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
fastapi==0.115.4
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SocketIO==5.4.1
gevent==24.10.3
gevent-websocket==0.10.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
iso8601==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pypika-tortoise==0.2.1
python-engineio==4.10.1
python-socketio==5.11.4
pytz==2024.2
setuptools==75.2.0
simple-websocket==1.1.0
sniffio==1.3.1
starlette==0.41.2
tortoise-orm==0.21.7
typing_extensions==4.12.2
uvicorn==0.32.0
Werkzeug==3.0.6
wsproto==1.2.0
zope.event==5.0
zope.interface==7.1.1
